+++ binding/Jsonnet.cpp
@@ -23,6 +23,8 @@ namespace nodejsonnet {
         InstanceMethod("evaluateSnippet", &Jsonnet::evaluateSnippet),
         InstanceMethod("evaluateFileMulti", &Jsonnet::evaluateFileMulti),
         InstanceMethod("evaluateSnippetMulti", &Jsonnet::evaluateSnippetMulti),
+        InstanceMethod("evaluateFileStream", &Jsonnet::evaluateFileStream),
+        InstanceMethod("evaluateSnippetStream", &Jsonnet::evaluateSnippetStream),
         InstanceMethod("extString", &Jsonnet::extString),
         InstanceMethod("extCode", &Jsonnet::extCode),
         InstanceMethod("tlaString", &Jsonnet::tlaString),
@@ -106,6 +108,29 @@ namespace nodejsonnet {
 
     auto vm = createVm(env);
     auto const worker = new JsonnetWorker(env, vm, std::make_unique<JsonnetWorker::EvaluateSnippetMultiOp>(std::move(snippet), std::move(filename)));
+    auto const promise = worker->Promise();
+    worker->Queue();
+    return promise;
+  }
+
+  Napi::Value Jsonnet::evaluateFileStream(const Napi::CallbackInfo& info) {
+    auto const env = info.Env();
+    auto filename = info[0].As<Napi::String>().Utf8Value();
+
+    auto vm = createVm(env);
+    auto const worker = new JsonnetWorker(env, vm, std::make_unique<JsonnetWorker::EvaluateFileStreamOp>(std::move(filename)));
+    auto const promise = worker->Promise();
+    worker->Queue();
+    return promise;
+  }
+
+  Napi::Value Jsonnet::evaluateSnippetStream(const Napi::CallbackInfo& info) {
+    auto const env = info.Env();
+    auto snippet = info[0].As<Napi::String>().Utf8Value();
+    auto filename = info.Length() < 2 ? "(snippet)" : info[1].As<Napi::String>().Utf8Value();
+
+    auto vm = createVm(env);
+    auto const worker = new JsonnetWorker(env, vm, std::make_unique<JsonnetWorker::EvaluateSnippetStreamOp>(std::move(snippet), std::move(filename)));
     auto const promise = worker->Promise();
     worker->Queue();
     return promise;