+++ binding/JsonnetVm.cpp
@@ -98,6 +98,24 @@ namespace nodejsonnet {
     return result;
   }
 
+  JsonnetVm::Buffer JsonnetVm::evaluateFileStream(std::string const &filename) const {
+    int error;
+    auto result = buffer(::jsonnet_evaluate_file_stream(vm, filename.c_str(), &error));
+    if(error != 0) {
+      throw std::runtime_error(result.get());
+    }
+    return result;
+  }
+
+  JsonnetVm::Buffer JsonnetVm::evaluateSnippetStream(std::string const &filename, std::string const &snippet) const {
+    int error;
+    auto result = buffer(::jsonnet_evaluate_snippet_stream(vm, filename.c_str(), snippet.c_str(), &error));
+    if(error != 0) {
+      throw std::runtime_error(result.get());
+    }
+    return result;
+  }
+
   JsonnetJsonValue *JsonnetVm::makeJsonString(std::string const &v) const {
     return ::jsonnet_json_make_string(vm, v.c_str());
   }