+++ binding/JsonnetVm.hpp
@@ -48,6 +48,8 @@ namespace nodejsonnet {
     Buffer evaluateSnippet(std::string const &filename, std::string const &snippet) const;
     Buffer evaluateFileMulti(std::string const &filename) const;
     Buffer evaluateSnippetMulti(std::string const &filename, std::string const &snippet) const;
+    Buffer evaluateFileStream(std::string const &filename) const;
+    Buffer evaluateSnippetStream(std::string const &filename, std::string const &snippet) const;
 
     JsonnetJsonValue *makeJsonString(std::string const &v) const;
     JsonnetJsonValue *makeJsonNumber(double v) const;