+++ types/index.d.ts
@@ -151,4 +151,21 @@ export class Jsonnet {
    * @returns A Promise for the evaluation result in JSON representation.
    */
   evaluateSnippetMulti(snippet: string, filename?: string): Promise<{[name: string]: string}>;
+
+  /**
+   * Evaluates a Jsonnet script in a file and returns a number of JSON's.
+   *
+   * @param filename - Path to the file.
+   * @returns A Promise for the evaluation result in JSON representation.
+   */
+  evaluateFileStream(filename: string): Promise<string[]>;
+
+  /**
+   * Evaluates a Jsonnet script given as a string and returns a number of JSON's.
+   *
+   * @param snippet - Jsonnet script to evaluate.
+   * @param filename - Filename used for error reporting.
+   * @returns A Promise for the evaluation result in JSON representation.
+   */
+  evaluateSnippetStream(snippet: string, filename?: string): Promise<string[]>;
 }